Added test for no test info found

+++ docs/examples/io_tests/execute.sh
@@ -223,7 +223,10 @@ then
 fi
 
 #rename the log file to accepted name called log.txt
-mv "$log" log.txt
+if [ -f "$log" ]
+then
+  mv "$log" log.txt
+fi
 
 #copy core dump messages to log file
 cp shellOut.txt temp.txt
